pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "inviter"
version = "0.1.0"
description = "Telegram Inviter Bot"
dependencies = [
    "pyrogram>=2.0.106",
    "TgCrypto>=1.2.5",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "httpx>=0.25.0",
    "aiosqlite>=0.19.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0"
]

[tool.setuptools.packages.find]
where = ["."]